djurable/{{cookiecutter.project_slug}}/_ws_config/asgi.py
import os
from channels.routing import ProtocolTypeRouter, URLRouter
from django.core.asgi import get_asgi_application
import django

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'config.settings')
django.setup()

from django.urls import path
from apps.core import consumers as core_consumers

application = ProtocolTypeRouter({
    'http': get_asgi_application(),
    'websocket': URLRouter([
        path('ws/example/', core_consumers.ExampleConsumer.as_asgi()),
    ]),
})
