djurable/pyproject.toml
[project]
name = "djurable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-unfold>=0.63.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "quixstreams>=3.22.0",
    "scrapy>=2.13.3",
]
